Add build scripts for client and server, update .gitignore, and enhance README.md
- Introduced multiple batch scripts for building and deploying client and server components, including `build_all_client.bat`, `build_libs.bat`, `build_engine.bat`, `build_client.bat`, `deploy_client.bat`, `build_servers.bat`, and `deploy_servers.bat`.
- Updated `.gitignore` to include the `upgrade` directory.
- Enhanced `README.md` with detailed migration notes and build instructions for the x64 architecture.

+++ sources/Client/src/GameLoading.cpp
@@ -189,12 +189,14 @@ void GameLoading::Close()
     GetClientRect(m_hLoading, &rc);
 
 
-    int height = rc.bottom - rc.top - (rc.right - rc.left) * 9.0 / 16.0;
+    // int height = rc.bottom - rc.top - (rc.right - rc.left) * 9.0 / 16.0;
+    double height = rc.bottom - rc.top - (rc.right - rc.left) * 9.0 / 16.0;
 
     RECT rect;
     rect.left = rc.left;
     rect.right = rc.right;
-    rect.top = rc.bottom - height / 2.0 - 1;
+    // rect.top = rc.bottom - height / 2.0 - 1;
+    rect.top = rc.bottom - (LONG)(height / 2.0) - 1;
     rect.bottom = rc.bottom;
 
     ::InvalidateRect(m_hLoading, &rect, TRUE);
